argument project 게시판 분리
issue게시판분리 , 불필요한 class 정리

+++ src/main/java/org/galapagos/domain/BoardVO.java
@@ -13,7 +13,7 @@ public class BoardVO {
 	
 	private Long bno;
 	
-	private int categoryId;
+	private Long categoryId;
 	
 	private String username;
